requirements.txt
pytest
requests
requests_mock
pytest-mock
python-dotenv